fix - #11: 로그인 기능 버그 수정
