+++ server/src/main/java/challenge/server/security/filter/JwtVerificationFilter.java
@@ -1,17 +1,30 @@
 package challenge.server.security.filter;
 
+import challenge.server.exception.BusinessLogicException;
+import challenge.server.exception.ExceptionCode;
+import challenge.server.security.jwt.JwtAuthenticationToken;
 import challenge.server.security.jwt.JwtTokenizer;
 import challenge.server.security.utils.CustomAuthorityUtils;
 import challenge.server.user.entity.LogoutList;
 import challenge.server.user.service.LogoutListService;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
+import io.jsonwebtoken.Jws;
 import io.jsonwebtoken.security.SignatureException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
@@ -21,69 +34,136 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
+import java.util.stream.Collectors;
 
 @Component
 @Slf4j
 @RequiredArgsConstructor
-public class JwtVerificationFilter /*extends OncePerRequestFilter*/ {
-//    private final JwtTokenizer jwtTokenizer;
-//    private final CustomAuthorityUtils authorityUtils;
-//    private final LogoutListService logoutListService;
-//
-//
-//    @Override
-//    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
-//        try {
-//            String jwt = resolveToken(request);
-//            String base64EncodedSecretKey = jwtTokenizer.encodeBase64SecretKey(jwtTokenizer.getSecretKey());
-//
-//            if (StringUtils.hasText(jwt) && jwtTokenizer.validateToken(jwt, base64EncodedSecretKey)) {
-//                LogoutList item = logoutListService.findByAccessToken(jwt);
-//                if (item == null) {
-//                    Map<String, Object> claims = verifyJws(request);
-//                    setAuthenticationToContext(claims);
-//                }
-//            }
-//        } catch (SignatureException se) {
-//            request.setAttribute("exception", se);
-//        } catch (ExpiredJwtException ee) {
-//            request.setAttribute("exception", ee);
-//        } catch (Exception e) {
-//            request.setAttribute("exception", e);
-//        }
-//
-//        filterChain.doFilter(request, response);
-//    }
-//
-//    @Override
-//    protected boolean shouldNotFilter(HttpServletRequest request) {
-//        String authorization = request.getHeader("Authorization");
-//        return authorization == null || !authorization.startsWith("Bearer");
-//    }
-//
-//    private Map<String, Object> verifyJws(HttpServletRequest request) {
-//        String jws = request.getHeader("Authorization").replace("Bearer ", "");
-//        String base64EncodedSecretKey = jwtTokenizer.encodeBase64SecretKey(jwtTokenizer.getSecretKey());
-//        Map<String, Object> claims = jwtTokenizer.getClaims(jws, base64EncodedSecretKey).getBody();
-//        return claims;
-//    }
-//
-//    private void setAuthenticationToContext(Map<String, Object> claims) {
-//        String username = (String) claims.get("username");
-//        List<GrantedAuthority> authorities = authorityUtils.createAuthorities((List) claims.get("roles"));
-//        Authentication authentication = new UsernamePasswordAuthenticationToken(username, null, authorities);
-//        SecurityContextHolder.getContext().setAuthentication(authentication);
-//    }
-//
-//    private String resolveToken(HttpServletRequest request) {
-//        String bearerToken = request.getHeader("Authorization");
-//
-//        if (StringUtils.hasText(bearerToken) && bearerToken.startsWith("Bearer")) {
-//            return bearerToken.substring(7); // 0~6번째 글자 제외한 문자열
-//        }
-//
-//        return null;
-//    }
+public class JwtVerificationFilter extends OncePerRequestFilter {
+    private final JwtTokenizer jwtTokenizer;
+    private final CustomAuthorityUtils authorityUtils;
+    private final LogoutListService logoutListService;
+
+
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
+        try {
+            String jwt = resolveToken(request);
+            String base64EncodedSecretKey = jwtTokenizer.encodeBase64SecretKey(jwtTokenizer.getSecretKey());
+
+            if (StringUtils.hasText(jwt) && jwtTokenizer.validateToken(jwt, base64EncodedSecretKey)) {
+                LogoutList item = logoutListService.findByAccessToken(jwt);
+                if (item == null) {
+                    Map<String, Object> claims = verifyJws(request);
+                    setAuthenticationToContext(claims);
+                }
+            }
+        } catch (SignatureException se) {
+            request.setAttribute("exception", se);
+        } catch (ExpiredJwtException ee) {
+            request.setAttribute("exception", ee);
+        } catch (Exception e) {
+            request.setAttribute("exception", e);
+        }
+
+        filterChain.doFilter(request, response);
+    }
+
+    @Override
+    protected boolean shouldNotFilter(HttpServletRequest request) {
+        String authorization = request.getHeader("Authorization");
+        return authorization == null || !authorization.startsWith("Bearer");
+    }
+
+    // JWT 토큰 복호화해서 토큰에 들어있는 정보 꺼냄
+    public Authentication getAuthentication(String accessToken, String base64EncodedSecretKey) {
+        // 토큰 복호화
+        Jws<Claims> claims = jwtTokenizer.getClaims(accessToken, base64EncodedSecretKey);
+
+        if (claims.getBody().getSubject() == null) { // 회원 email
+            throw new BusinessLogicException(ExceptionCode.AUTHENTICATION_NOT_FOUND);
+        }
+
+        // 클레임에서 권한 정보 가져오기
+        Collection<? extends GrantedAuthority> authorities =
+                Arrays.stream(claims.getBody().getSubject().toString().split(","))
+                        .map(SimpleGrantedAuthority::new)
+                        .collect(Collectors.toList());
+
+        // UserDetails 객체를 만들어서 Authentication 리턴
+        UserDetails principal = new User(claims.getBody().getSubject(), "", authorities);
+        return new UsernamePasswordAuthenticationToken(principal, "", authorities);
+    }
+
+    private Map<String, Object> verifyJws(HttpServletRequest request) {
+        String jws = request.getHeader("Authorization").replace("Bearer ", "");
+        String base64EncodedSecretKey = jwtTokenizer.encodeBase64SecretKey(jwtTokenizer.getSecretKey());
+        Map<String, Object> claims = jwtTokenizer.getClaims(jws, base64EncodedSecretKey).getBody();
+        return claims;
+    }
+
+    private void setAuthenticationToContext(Map<String, Object> claims) {
+        String username = (String) claims.get("username");
+        List<GrantedAuthority> authorities = authorityUtils.createAuthorities((List) claims.get("roles"));
+        Authentication authentication = new UsernamePasswordAuthenticationToken(username, null, authorities);
+        SecurityContextHolder.getContext().setAuthentication(authentication);
+    }
+
+    private String resolveToken(HttpServletRequest request) {
+        String bearerToken = request.getHeader("Authorization");
+
+        if (StringUtils.hasText(bearerToken) && bearerToken.startsWith("Bearer")) {
+            return bearerToken.substring(7); // 0~6번째 글자 제외한 문자열
+        }
+
+        return null;
+    }
+
+    private String delegateAccessToken(challenge.server.user.entity.User user) {
+        Map<String, Object> claims = new HashMap<>();
+//        claims.put("userId", user.getUserId());
+        claims.put("username", user.getUsername());
+        claims.put("roles", user.getRoles());
+
+        String subject = user.getEmail();
+        Date expiration = jwtTokenizer.getTokenExpiration(jwtTokenizer.getAccessTokenExpirationMinutes());
+        String base64EncodedSecretKey = jwtTokenizer.encodeBase64SecretKey(jwtTokenizer.getSecretKey());
+        String accessToken = jwtTokenizer.generateAccessToken(claims, subject, expiration, base64EncodedSecretKey);
+        return accessToken;
+    }
+
+    private String delegateRefreshToken(challenge.server.user.entity.User user) {
+        String subject = user.getEmail();
+        Date expiration = jwtTokenizer.getTokenExpiration((jwtTokenizer.getRefreshTokenExpirationMinutes()));
+        String base64EncodedSecretKey = jwtTokenizer.encodeBase64SecretKey(jwtTokenizer.getSecretKey());
+        String refreshToken = jwtTokenizer.generateRefreshToken(subject, expiration, base64EncodedSecretKey);
+        return refreshToken;
+    }
+
+    public String reissueRefreshToken(challenge.server.user.entity.User user, HttpServletResponse response) throws ServletException, IOException {
+        String refreshToken = delegateRefreshToken(user);
+
+        response.setHeader("Refresh", refreshToken);
+
+        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+        response.setStatus(HttpStatus.OK.value());
+
+//        Map<String, Long> body = new HashMap<>();
+//        body.put("userId", user.getUserId()); // 로그인 후 userId를 응답 body에 전달
+
+//        new ObjectMapper().writeValue(response.getOutputStream(), body);
+
+        return refreshToken;
+    }
+
+    public String reissueAccessToken(challenge.server.user.entity.User user, HttpServletResponse response) {
+        String accessToken = delegateAccessToken(user);
+        response.setHeader("Authorization", "Bearer " + accessToken);
+
+        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+        response.setStatus(HttpStatus.OK.value());
+
+        return accessToken;
+    }
 }
