+++ server/src/main/java/challenge/server/security/jwt/JwtTokenizer.java
@@ -10,10 +10,8 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Component;
 
-import javax.servlet.http.HttpServletRequest;
 import java.nio.charset.StandardCharsets;
 import java.security.Key;
-import java.time.Instant;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.Map;
@@ -60,9 +58,6 @@ public class JwtTokenizer {
                 .compact();
     }
 
-    /**
-     * jwt token 안에 있는 exp를 구하기 위한 메서드
-     */
     public Jws<Claims> getClaims(String jws, String base64EncodedSecretKey) {
         Key key = getKeyFromBase64EncodedKey(base64EncodedSecretKey);
 
@@ -72,34 +67,6 @@ public class JwtTokenizer {
                 .parseClaimsJws(jws);
 
         return claims;
-    }
-
-    public Date extractDateFromToken(String jwtToken) {
-        Jws<Claims> claims = getClaims(jwtToken, makeBase64EncodedSecretKey(secretKey));
-        int exp = (int) claims.getBody().get("exp"); // todo "exp"가 뭐지?
-        return Date.from(Instant.ofEpochSecond(exp));
-    }
-
-    /**
-     * 회원 유효성 검증
-     * @param request -> 헤더값 Authorization 추출
-     * @return 맴버의 이메일
-     */
-    public String getEmailFromToken(HttpServletRequest request) {
-        // header에 있는, 복호화된? 값을 가져옴
-        String authorization = request.getHeader("Authorization");
-
-        try {
-            Claims body = Jwts.parserBuilder()
-                    .setSigningKey(getKeyFromBase64EncodedKey(makeBase64EncodedSecretKey(secretKey)))
-                    .build()
-                    .parseClaimsJws(authorization)
-                    .getBody(); // 값을 넣어서 되돌려받는다(payload 값만)
-
-            return (String) body.get("username"); // todo "sub"이 뭐지? -> 이메일 반환
-        } catch (Exception e) {
-            throw e;
-        }
     }
 
     public void verifySignature(String jws, String base64EncodedSecretKey) {
@@ -119,18 +86,6 @@ public class JwtTokenizer {
         return expiration;
     }
 
-    /**
-     * 1번
-     * 보안키 비번을 암호화
-     */
-    public String makeBase64EncodedSecretKey(String secretKey) {
-        return Encoders.BASE64.encode(secretKey.getBytes(StandardCharsets.UTF_8));
-    }
-
-    /**
-     * 2번
-     * 비밀번호 제작 메서드
-     */
     private Key getKeyFromBase64EncodedKey(String base64EncodedSecretKey) {
         byte[] keyBytes = Decoders.BASE64.decode(base64EncodedSecretKey);
         Key key = Keys.hmacShaKeyFor(keyBytes);
@@ -142,16 +97,8 @@ public class JwtTokenizer {
         Key key = getKeyFromBase64EncodedKey(base64EncodedSecretKey);
 
         try {
-            // v1
-//            Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token);
-//            return true;
-
-            // v2
-            Jws<Claims> claimsJws = Jwts.parserBuilder()
-                    .setSigningKey(key)
-                    .build()
-                    .parseClaimsJws(token);
-            return !claimsJws.getBody().getExpiration().before(new Date());
+            Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token);
+            return true;
         } catch (io.jsonwebtoken.security.SecurityException | MalformedJwtException e) {
             log.info("Invalid JWT token", e);
         } catch (ExpiredJwtException e) {
