--- server/src/main/java/challenge/server/security/userdetails/CustomUserDetailsService2.java
@@ -1,41 +1,0 @@
-package challenge.server.security.userdetails;
-
-import challenge.server.user.entity.User;
-import challenge.server.user.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Component("userDetailsService")
-@RequiredArgsConstructor
-public class CustomUserDetailsService2 implements UserDetailsService {
-    private final UserRepository userRepository;
-
-    @Override
-    @Transactional
-    public UserDetails loadUserByUsername(final String username) {
-        return userRepository.findByEmail(username)
-                .map(user -> createUser(username, user))
-                .orElseThrow(() -> new UsernameNotFoundException(username + " -> 데이터베이스에서 찾을 수 없습니다."));
-    }
-
-    private org.springframework.security.core.userdetails.User createUser(String username, User user) {
-        if (user.getStatus() != User.Status.ACTIVE) {
-            throw new RuntimeException(username + " -> 활성화되어 있지 않습니다.");
-        }
-        List<GrantedAuthority> grantedAuthorities = user.getRoles().stream()
-                .map(role -> new SimpleGrantedAuthority(role))
-                .collect(Collectors.toList());
-        return new org.springframework.security.core.userdetails.User(user.getUsername(),
-                user.getPassword(),
-                grantedAuthorities);
-    }
-}
