+++ server/src/main/java/challenge/server/user/controller/RefreshTokenController.java
@@ -1,7 +1,7 @@
 package challenge.server.user.controller;
 
-import challenge.server.security.service.SecurityService;
 import challenge.server.user.dto.UserDto;
+import challenge.server.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -23,11 +23,11 @@ import java.io.IOException;
 @Slf4j
 @RequiredArgsConstructor
 public class RefreshTokenController {
-    private final SecurityService securityService;
+    private final UserService userService;
 
     @PostMapping
     public ResponseEntity reissueToken(@Valid @RequestBody UserDto.TokenRequest requestBody, HttpServletResponse response) throws ServletException, IOException {
-        securityService.reissueToken(requestBody, response);
+        userService.reissueToken(requestBody, response);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 }
